+++ openair2/RRC/LITE/MESSAGES/asn1_msg.h
@@ -212,7 +212,7 @@ uint8_t do_MBSFNAreaConfig(uint8_t Mod_id,
                            MBSFNAreaConfiguration_r9_t **mbsfnAreaConfiguration);
 #endif 
 
-uint8_t do_MeasurementReport(uint8_t Mod_id, uint8_t *buffer,int measid,int phy_id,int rsrp_s,int rsrq_s,long rsrp_t,long rsrq_t);
+uint8_t do_MeasurementReport(uint8_t Mod_id, uint8_t *buffer,int measid,int phy_id,long rsrp_s,long rsrq_s,long rsrp_t,long rsrq_t);
 
 uint8_t do_DLInformationTransfer(uint8_t Mod_id, uint8_t **buffer, uint8_t transaction_id, uint32_t pdu_length, uint8_t *pdu_buffer);
